API-31: Various java.util.Date subclasses can break DateTypeSupport
Rather then cloning the original Date object, we create new instance
based on the original Date's object epoch. This way we ensure that the
object inside Record is always java.util.Date and not any custom
sub-class that would change behavior of toString() and other methods
that important for our serialization code.

Change-Id: Ic99a2eac94673adbc7e98b5dd60e73736829450b
Reviewed-on: https://review.streamsets.net/3322
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>