SDC-4470. Add methods to get and replace all header attributes in a record
Kryo deserializes the Records in a different classloader. This makes it impossible
to use the RecordImpl objects from Spark directly.

These methods allow us to copy
the original header data into the new records which are created using the original one
as its originator. This way we don't lose stage paths etc.

Change-Id: I72c98be3ea603d981167d76f7250ac640fa75490
Reviewed-on: https://review.streamsets.net/5230
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>