API-73: Add method toEvent and toError to PushSource.Context
Adding support for events and errors to push source is a bit tricky
because the StageContext is shared for all threads. I could think about
two solutions:

1) We could convert the sinks to be ThreadLocal which would allow
us to differentiate to which batch each of the sinks belongs to. The
limitation is that each batch then have to be process by exactly one
thread. Calling processBatch from a different thread would not work and
could potentially cause an unexpected loss of events and error records.

2) For push origin, we could enhance the BatchContext class to provide
methods toEvent and toError. Even though this is sort of a different
way, it puts burden on doing everything correctly on the user rather
then us and thus making the framework contract easier to see and follow.

This patch implements API part of 2).

Change-Id: I9aa47f7a98f4d639af7a1de9bcf930e03f7c954d
Reviewed-on: https://review.streamsets.net/5786
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>