API-142: CredentialsValue doesn't work properly in complex types
This happened after we transitioned the component to credentials
framework. Just placing any component using Hikari config will lead to
validation error about not allowing nulls in JDBC Driver properties.

It turns out that our code is written in a way that requires the default
value to not be null (otherwise it's an error), but we mistakenly used
null as default value for CREDENTIAL.

This patch sets default value to empty string which matches the expected
behavior of empty credentials object.

Change-Id: I267c7ee89dd3d2367f42620a7b8197fcf5b28781
Reviewed-on: https://review.streamsets.net/9775
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>