API-169: Extract ConfigIssue to be shared between Service and Stage
This patch extracts ConfigIssue out of Stage and makes it reusable. The
change is written in backward compatible way for existing Stages - e.g.
Stage implementation should not need any change and still compile with
the new change. Service implementation will however need to be changed,
but that is fine since it's still under active development.

Change-Id: I16c4b157baba32330a21a13f539e20e905a512db
Reviewed-on: https://review.streamsets.net/12162
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>