API-170: Service: Create shared parent class for both Service and Stage
I was hoping that I'll be able to avoid creating a parent interface for
both Stage and Service, but it seems that it's actually the best way how
to share some methods and context classes.

This patch is intentionally minimal - it's meant as a first step that we
will iterate on later (e.g. migrate more context methods or even migrate
the init() and destroy()).

Change-Id: Ia1bbc6b02eecd9cf73078faaee0a04f3fa67ef83
Reviewed-on: https://review.streamsets.net/12163
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>