API-176: Service: Migrate StageUpgrader to use Upgrade Context
This patch changes the way we expect upgrader to be used - rather then
having a single method that takes a long list of parameters to which
it's difficult to add a new one, it converts the method to use a Context
object. We can at any time add a new method to this Context without
breaking existing implementations. This is what we already do in other
components that user can implement (Stage, Service, ....).

The change is done in backward compatible manner - e.g. the new method
have default implementation that calls the older implementation.

The new context also contains newly introduce method registerService()
that will enable upgrader to inject a configuration for a service when
migrating from pre-service to a service implementation of the stage.

Change-Id: I00f96619ff62bb7a3047dd7ba3a4aa58e5687af3
Reviewed-on: https://review.streamsets.net/12206
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>