API-194: Service: Introduce DStage(s) to API module
This patch migrates the existing DStage(s) from commonlib to API so that
stages outside of our own code base can use that. This does include
stages that even we build, but that lives outside of data collector code
base. I'm introducing it for services, but true to be told, I do need
this capability in other cases as well.

Change-Id: I82151873e8aa40b7736cb1bbb4d6f155922f0bad
Reviewed-on: https://review.streamsets.net/13853
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>