API-224: Introduce HideStage annotation
The idea is to use a special annotation for stages that should
not be visible on the main pipeline canvas. The annotation has
a parameter instructing the framework where else is the stage
allowed. This initial version only supports field processor -
transitioning error and stats stage will happen separately and
in the future change.

Change-Id: Ib0868f09d776379f6debdd55a9aaa378b814d06b
Reviewed-on: https://review.streamsets.net/14985
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>