API-228: BaseFieldProcessor requires at least one output lane
I've originally implemented the BaseFieldProcessor as inheriting from
SingleLaneRecordProcessor, but that actually doesn't make sense as this
processor requires (rightfully so) exactly 1 output lane. However since
the FieldProcessor will be mostly used in cases where there are no
output lanes, the inheritance doesn't make much sense and hence I
changed it.

Since the number of output lanes is not known now, I've changed the
default implementation of process() to thrown an exception and either
force the subclass to properly re-implemented the process() or only
depend on the process(FieldBatch).

Change-Id: Idb7d2269d11359bf62814af3ca781fbf121a774a
Reviewed-on: https://review.streamsets.net/15060
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>