API-324: Re-implement TextUtils.isValidName
It turns out that rather then using regular expression, simple iteration
over the name and validating character classes is much much faster (on
my weird and old machine in the range of 5-7x). This will help with
pipelines relying a lot on expression evaluator as ELVars rely on valid
names everywhere.

Change-Id: I2402b4b743bdeab0ec9e8c8a781539bb98b053c0
Reviewed-on: https://review.streamsets.net/c/24151
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>